+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/class_loading/Boot.java
@@ -1,22 +1,27 @@
 package dev.kostromdan.mods.crash_assistant.app.class_loading;
 
-import dev.kostromdan.mods.crash_assistant.loading_utils.JavaBinaryLocator;
+import dev.kostromdan.mods.crash_assistant.common_config.loading_utils.JavaBinaryLocator;
 
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.Base64;
 import java.util.List;
+import java.util.Objects;
 
 public class Boot {
     public static String log4jApi = null;
     public static String log4jCore = null;
     public static String googleGson = null;
     public static String commonIo = null;
+    public static String lwjglNatives = null;
     public static String processor = null;
     public static String jarPath = null;
     public static boolean recursiveStart = false;
+    public static String serialisedGPUs = null;
     public static List<String> JVM_ARGS = ManagementFactory.getRuntimeMXBean().getInputArguments();
     public static List<String> APP_ARGS;
 
@@ -32,10 +37,14 @@ public class Boot {
                 googleGson = args[i + 1];
             } else if ("-commonIo".equals(args[i]) && i + 1 < args.length) {
                 commonIo = args[i + 1];
+            } else if ("-lwjglNatives".equals(args[i]) && i + 1 < args.length) {
+                lwjglNatives = args[i + 1];
             } else if ("-processor".equals(args[i]) && i + 1 < args.length) {
                 processor = args[i + 1];
             } else if ("-jarPath".equals(args[i]) && i + 1 < args.length) {
                 jarPath = args[i + 1];
+            } else if ("-serialisedGPUs".equals(args[i]) && i + 1 < args.length) {
+                serialisedGPUs = new String(Base64.getDecoder().decode(args[i + 1]), StandardCharsets.UTF_8);
             } else if ("-recursiveStart".equals(args[i])) {
                 recursiveStart = true;
             }
@@ -48,27 +57,50 @@ public class Boot {
             System.exit(-1);
         }
 
+        CrashAssistantAgent.appendJarFile(log4jApi);
+        CrashAssistantAgent.appendJarFile(log4jCore);
+//        CrashAssistantAgent.appendJarFile(googleGson);
+        CrashAssistantAgent.appendJarFile(commonIo);
+        if (lwjglNatives != null && !Objects.equals(lwjglNatives, "UNDEFINED") && !recursiveStart) {
+            CrashAssistantAgent.appendJarFile(lwjglNatives);
+        } else {
+            lwjglNatives = null;
+        }
+
         /**
          * If Minecraft JVM terminated by windows itself, all child processes will be also terminated.
          * So Crash Assistant can't be child process. This way we make Crash Assistant completely independent process.
+         *
+         * Also, here we're locating GPUs with Vulkan, it's increasing heap before GUI start,
+         * so we're doing it on this TMP process, to not waste user resources on App avaiting stage.
          */
         if (!recursiveStart) {
+            if (log4jApi != null) {
+                try {
+                    Class<?> crashAssistantAppClass = Class.forName("dev.kostromdan.mods.crash_assistant.app.utils.gpu.GPUDetector");
+                    Method mainMethod = crashAssistantAppClass.getMethod("getSerialisedGPUs");
+                    serialisedGPUs = (String) mainMethod.invoke(null);
+                } catch (Exception e) {
+                    e.printStackTrace(System.err);
+                }
+            }
+
             List<String> argsList = new ArrayList<>();
             argsList.add(JavaBinaryLocator.getJavaBinary(ProcessHandle.current()));
             argsList.addAll(JVM_ARGS);
             argsList.add("-jar");
             argsList.add(jarPath);
             argsList.addAll(APP_ARGS);
+            if (serialisedGPUs != null) {
+                String encodedGPUs = Base64.getEncoder().encodeToString(serialisedGPUs.getBytes(StandardCharsets.UTF_8));
+                argsList.add("-serialisedGPUs");
+                argsList.add(encodedGPUs);
+            }
             argsList.add("-recursiveStart");
             ProcessBuilder pb = new ProcessBuilder(argsList);
             pb.start();
             System.exit(0);
         }
-
-        CrashAssistantAgent.appendJarFile(log4jApi);
-        CrashAssistantAgent.appendJarFile(log4jCore);
-        CrashAssistantAgent.appendJarFile(googleGson);
-        CrashAssistantAgent.appendJarFile(commonIo);
 
         Class<?> crashAssistantAppClass = Class.forName("dev.kostromdan.mods.crash_assistant.app.CrashAssistantApp");
         Method mainMethod = crashAssistantAppClass.getMethod("main", String[].class);