+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/logs_analyser/LogAnalyser.java
@@ -7,9 +7,8 @@ import dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.hs_er
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.log.*;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.log.OutOfMemoryError;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.win_event.WasClosedByWindows;
-import dev.kostromdan.mods.crash_assistant.config.CrashAssistantConfig;
+import dev.kostromdan.mods.crash_assistant.common_config.config.CrashAssistantConfig;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -20,6 +19,7 @@ import java.util.concurrent.TimeUnit;
 public class LogAnalyser {
     private static final List<KnownCrashReason> registeredReasons = new ArrayList<>();
     private static final List<CodexAnalysis> registeredCodexReasons = new ArrayList<>();
+    private static boolean reasonsRegistered = false;
     public static final HashSet<LogType> CodexSupportedLogTypes = new HashSet<>() {{
         add(LogType.LOG);
         add(LogType.CRASH_REPORT);
@@ -37,9 +37,16 @@ public class LogAnalyser {
         if (!CrashAssistantConfig.getBoolean("analysis.enabled")) {
             return;
         }
-        ExecutorService pool = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors());
-        HashSet<String> disabledCrashReasons = new HashSet<>(CrashAssistantConfig.getBlacklistedAnalysis());
+        long startTime = System.currentTimeMillis();
+        registerReasons();
         synchronized (KnownCrashReasonMessage.class) {
+            ExecutorService pool = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors());
+            HashSet<String> disabledCrashReasons = new HashSet<>(CrashAssistantConfig.getBlacklistedAnalysis());
+            for (Log log : LogsList.getLogs()) {
+                pool.submit(() -> {
+                    log.getProcessor().processLogFileSafe();
+                });
+            }
             for (Log log : LogsList.getLogs()) {
                 analyseLog(log, disabledCrashReasons, pool);
             }
@@ -53,19 +60,14 @@ public class LogAnalyser {
                 Thread.currentThread().interrupt();
                 CrashAssistantApp.LOGGER.error("Interrupted while awaiting termination of analysis tasks", e);
             }
+            CrashAssistantApp.LOGGER.info("Analysis finished in {} ms", System.currentTimeMillis() - startTime);
         }
     }
 
 
-    public static synchronized void analyseLog(Log log, HashSet<String> disabledCrashReasons, ExecutorService pool) {
+    private static synchronized void analyseLog(Log log, HashSet<String> disabledCrashReasons, ExecutorService pool) {
         if (log.isAnalysed()) return;
 
-        try {
-            log.getProcessor().processLogFile();
-        } catch (IOException e) {
-            CrashAssistantApp.LOGGER.error("Error processing log file", e);
-        }
-        String logText = log.getProcessor().getAllLinesString();
         List<KnownCrashReason> registeredReasonsForThisLog = registeredReasons.stream()
                 .filter(reason ->
                         reason.getLogTypes().contains(log.getType()) &&
@@ -74,7 +76,11 @@ public class LogAnalyser {
 
         for (KnownCrashReason reason : registeredReasonsForThisLog) {
             pool.submit(() -> {
-                if (reason.matches(logText, log)) {
+                log.getProcessor().processLogFileSafe();
+                String logText = log.getProcessor().getAllLinesString();
+                if (reason.matches(logText, log)
+//                        || true //dubug too see all available warnings
+                ) {
                     synchronized (pool) {
                         KnownCrashReasonMessage.addCrashReasonMessage(
                                 new KnownCrashReasonMessage(log, reason)
@@ -87,6 +93,7 @@ public class LogAnalyser {
     }
 
     public static synchronized String analyseCodexMessage(String message) {
+        registerReasons();
         for (CodexAnalysis reason : registeredCodexReasons) {
             if (reason.matches(message)) {
                 return reason.getMessage();
@@ -95,7 +102,10 @@ public class LogAnalyser {
         return "";
     }
 
-    public static void registerReasons() {
+    public static synchronized void registerReasons() {
+        if (reasonsRegistered) {
+            return;
+        }
         registerKnownCrashReason(new Atio6axx());
         registerKnownCrashReason(new InsufficientMemory());
         registerKnownCrashReason(new Jemalloc());
@@ -106,6 +116,7 @@ public class LogAnalyser {
         registerKnownCrashReason(new Create6Addons());
         registerKnownCrashReason(new CtovWithoutLithostitched());
         registerKnownCrashReason(new CurseForgeCorrupted());
+        registerKnownCrashReason(new DiskSpaceEnded());
         registerKnownCrashReason(new DuplicatedMods());
         registerKnownCrashReason(new MissingEmbeddiumForOculus());
         registerKnownCrashReason(new OutOfMemoryError());
@@ -115,5 +126,7 @@ public class LogAnalyser {
 
 
         registerCodexKnownCrashReason(new ErroringEntity());
+
+        reasonsRegistered = true;
     }
 }