+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/logs_analyser/RegexChecker.java
@@ -7,26 +7,6 @@ import java.util.Collection;
 import java.util.regex.Pattern;
 
 public class RegexChecker {
-    public static boolean logContainsOneOfPatterns(Log log, Collection<String> patterns) {
-        if (patterns == null) {
-            return false;
-        }
-        return logContainsOneOfPatterns(log, patterns.toArray(new String[0]));
-    }
-
-    public static boolean logContainsOneOfPatterns(Log log, String... patterns) {
-        if (patterns.length == 0 || !log.getFile().isFile()) {
-            return false;
-        }
-        try {
-            log.getProcessor().processLogFile();
-            return logContainsOneOfPatterns(log.getProcessor().getAllLinesString(), log.getPath(), patterns);
-        } catch (Exception e) {
-            CrashAssistantApp.LOGGER.error("Error while reading " + log.getFileName() + " file: ", e);
-            return false;
-        }
-    }
-
     public static boolean logContainsOneOfPatterns(String logContents, Path logFile, Collection<String> patterns) {
         if (patterns == null) {
             return false;