+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/logs_analyser/crash_reasons/hs_err/Atio6axx.java
@@ -2,7 +2,7 @@ package dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.hs_e
 
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.KnownCrashReason;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.LogType;
-import dev.kostromdan.mods.crash_assistant.lang.LanguageProvider;
+import dev.kostromdan.mods.crash_assistant.common_config.lang.LanguageProvider;
 
 public class Atio6axx extends KnownCrashReason {
     public Atio6axx() {