+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/logs_analyser/crash_reasons/log/Create6Addons.java
@@ -5,13 +5,16 @@ import dev.kostromdan.mods.crash_assistant.app.gui.CreateDependencies;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.KnownCrashReason;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.Log;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.LogType;
-import dev.kostromdan.mods.crash_assistant.lang.LanguageProvider;
-import dev.kostromdan.mods.crash_assistant.mod_list.Mod;
-import dev.kostromdan.mods.crash_assistant.mod_list.ModListDiff;
-import dev.kostromdan.mods.crash_assistant.mod_list.ModListUtils;
-import dev.kostromdan.mods.crash_assistant.platform.PlatformHelp;
+import dev.kostromdan.mods.crash_assistant.common_config.lang.LanguageProvider;
+import dev.kostromdan.mods.crash_assistant.common_config.mod_list.Mod;
+import dev.kostromdan.mods.crash_assistant.common_config.mod_list.ModListDiff;
+import dev.kostromdan.mods.crash_assistant.common_config.mod_list.ModListUtils;
+import dev.kostromdan.mods.crash_assistant.common_config.platform.PlatformHelp;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
 
 
 public class Create6Addons extends KnownCrashReason {