+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/logs_analyser/crash_reasons/log/CurseForgeCorrupted.java
@@ -4,7 +4,7 @@ import dev.kostromdan.mods.crash_assistant.app.logs_analyser.KnownCrashReason;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.Log;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.LogType;
 import dev.kostromdan.mods.crash_assistant.app.utils.FileUtils;
-import dev.kostromdan.mods.crash_assistant.lang.LanguageProvider;
+import dev.kostromdan.mods.crash_assistant.common_config.lang.LanguageProvider;
 
 import java.util.HashMap;
 
@@ -17,7 +17,7 @@ public class CurseForgeCorrupted extends KnownCrashReason {
                 }}),
                 "Missing or unsupported mandatory dependencies:\\R\\s*Mod ID: 'minecraft', Requested by: .*?, Actual version: '\\[MISSING\\]'",
                 "Missing or unsupported mandatory dependencies:\\R\\s*Mod ID: 'neoforge', Requested by: .*?, Actual version: '\\[MISSING\\]'",
-                "Error loading class: net/minecraft/world/level/biome/Biome \\(java\\.lang\\.IllegalStateException: Unable to find method \\(\\)Lnet/minecraft/world/level/biome/BiomeSpecialEffects; with name getModifiedSpecialEffects in net/minecraft/world/level/biome/Biome\\)"
+                "Error loading class: net/minecraft/world/level/biome/Biome \\(java\\.lang\\.IllegalStateException: Unable to find method \\(\\)Lnet/minecraft/world/level/biome/"
         );
     }