+++ app/src/main/java/dev/kostromdan/mods/crash_assistant/app/logs_analyser/crash_reasons/log/DuplicatedMods.java
@@ -1,20 +1,11 @@
 package dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.log;
 
 import dev.kostromdan.mods.crash_assistant.app.CrashAssistantApp;
-import dev.kostromdan.mods.crash_assistant.app.gui.CreateDependencies;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.KnownCrashReason;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.Log;
 import dev.kostromdan.mods.crash_assistant.app.logs_analyser.LogType;
-import dev.kostromdan.mods.crash_assistant.app.logs_analyser.crash_reasons.hs_err.InsufficientMemory;
-import dev.kostromdan.mods.crash_assistant.lang.LanguageProvider;
-import dev.kostromdan.mods.crash_assistant.mod_list.Mod;
-import dev.kostromdan.mods.crash_assistant.mod_list.ModListDiff;
-import dev.kostromdan.mods.crash_assistant.mod_list.ModListUtils;
-import dev.kostromdan.mods.crash_assistant.platform.PlatformHelp;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Objects;
+import dev.kostromdan.mods.crash_assistant.common_config.lang.LanguageProvider;
+import dev.kostromdan.mods.crash_assistant.common_config.platform.PlatformHelp;
 
 public class DuplicatedMods extends KnownCrashReason {
     public DuplicatedMods() {