+++ common_config/src/main/java/dev/kostromdan/mods/crash_assistant/common_config/loading_utils/JarLocatingException.java
@@ -1,0 +1,7 @@
+package dev.kostromdan.mods.crash_assistant.common_config.loading_utils;
+
+public class JarLocatingException extends RuntimeException {
+    public JarLocatingException(String message) {
+        super(message);
+    }
+}